+++ resources/js/wysiwyg/lexical/core/nodes/LexicalRootNode.ts
@@ -99,8 +99,6 @@ export class RootNode extends ElementNode {
   static importJSON(serializedNode: SerializedRootNode): RootNode {
     // We don't create a root, and instead use the existing root.
     const node = $getRoot();
-    node.setFormat(serializedNode.format);
-    node.setIndent(serializedNode.indent);
     node.setDirection(serializedNode.direction);
     return node;
   }
@@ -109,8 +107,6 @@ export class RootNode extends ElementNode {
     return {
       children: [],
       direction: this.getDirection(),
-      format: this.getFormatType(),
-      indent: this.getIndent(),
       type: 'root',
       version: 1,
     };
